moved bill embedding function to separate script. Scraper now sends e_embed_bills message
