--- test_congress_scraper.py
@@ -1,157 +1,0 @@
-#!/usr/bin/env python3
-"""
-Test script for congress_scraper semantic filtering functionality.
-This script tests the new semantic filtering approach that resolves the issue
-where different queries were returning the same bills.
-"""
-
-import os
-import sys
-
-# Add the scraper logic directory to the path
-sys.path.append('src/scraper-lambda/logic')
-
-# Set environment variables (replace with your actual API key)
-os.environ['CONGRESS_API_KEY'] = '3Dii9BEerep10C3SCebdeYxJrdnTEmexYrKQaeLo'
-
-from congress_scraper import Congress
-
-def test_semantic_filtering():
-    """Test semantic filtering with different interest categories."""
-    print('Testing Congress scraper with semantic filtering:')
-    print('=' * 60)
-
-    # Test 1: Healthcare-related interests
-    healthcare_interests = ['healthcare', 'medical research', 'prescription drugs']
-    print(f'\n1. Testing with healthcare interests: {healthcare_interests}')
-    congress_health = Congress(healthcare_interests)
-    health_bills = congress_health.get_bills()
-
-    if health_bills:
-        print(f'   Found {len(health_bills)} bills matching healthcare interests')
-        for i, bill in enumerate(health_bills[:3]):
-            print(f'   {i+1}. {bill["title"][:70]}...')
-            print(f'      Similarity: {bill.get("similarity_score", "N/A"):.3f} (Matched: {bill.get("keyword", "N/A")})')
-    else:
-        print('   No healthcare bills found')
-
-    # Test 2: Technology-related interests  
-    tech_interests = ['artificial intelligence', 'cybersecurity', 'technology innovation']
-    print(f'\n2. Testing with technology interests: {tech_interests}')
-    congress_tech = Congress(tech_interests)
-    tech_bills = congress_tech.get_bills()
-
-    if tech_bills:
-        print(f'   Found {len(tech_bills)} bills matching technology interests')
-        for i, bill in enumerate(tech_bills[:3]):
-            print(f'   {i+1}. {bill["title"][:70]}...')
-            print(f'      Similarity: {bill.get("similarity_score", "N/A"):.3f} (Matched: {bill.get("keyword", "N/A")})')
-    else:
-        print('   No technology bills found')
-
-    # Compare results
-    print('\n' + '=' * 60)
-    print('COMPARISON RESULTS:')
-    if health_bills and tech_bills:
-        health_titles = {bill['title'] for bill in health_bills}
-        tech_titles = {bill['title'] for bill in tech_bills}
-        overlap = health_titles.intersection(tech_titles)
-        
-        print(f'Healthcare bills: {len(health_bills)}')
-        print(f'Technology bills: {len(tech_bills)}')
-        print(f'Overlapping bills: {len(overlap)}')
-        
-        if len(overlap) < min(len(health_bills), len(tech_bills)):
-            print('✅ SUCCESS: Different interests are returning different bills!')
-        else:
-            print('❌ ISSUE: Too much overlap between different interest categories')
-    else:
-        print('One or both categories returned no results - this may be normal')
-
-def test_broad_interests():
-    """Test with broader interests and lower threshold to see more results."""
-    print('\n\nTesting with broader interests and lower threshold:')
-    print('=' * 60)
-
-    # Create Congress instance with broader interests
-    broad_interests = ['trade', 'economics', 'cybersecurity', 'veterans', 'education', 'infrastructure']
-    print(f'Testing with interests: {broad_interests}')
-
-    congress = Congress(broad_interests)
-    # Temporarily lower the threshold to see more results
-    congress.similarity_threshold = 0.25  # Lower threshold
-    print(f'Using similarity threshold: {congress.similarity_threshold}')
-
-    bills = congress.get_bills()
-
-    if bills:
-        print(f'\nFound {len(bills)} bills matching interests')
-        print('\nTop 10 bills by similarity score:')
-        for i, bill in enumerate(bills[:10]):
-            similarity = bill.get('similarity_score', 'N/A')
-            keyword = bill.get('keyword', 'N/A')
-            print(f'  {i+1}. {bill["title"][:70]}...')
-            print(f'     Similarity: {similarity:.3f} | Matched Interest: "{keyword}"')
-            print()
-    else:
-        print('No bills found even with lower threshold')
-
-def test_handler_function():
-    """Test the Lambda handler function."""
-    print('\n\nTesting Lambda handler function:')
-    print('=' * 60)
-    
-    from congress_scraper import handler
-    
-    # Test payload
-    test_payload = {
-        "topics": ["cybersecurity", "veterans", "education"]
-    }
-    
-    print(f'Testing handler with payload: {test_payload}')
-    result = handler(test_payload)
-    
-    print(f'\nHandler returned status code: {result["statusCode"]}')
-    if result["statusCode"] == 200:
-        import json
-        response_data = json.loads(result["body"])
-        print(f'Handler response: {response_data.get("message", "Success")}')
-    else:
-        print(f'Handler response: {result["body"]}')
-
-def main():
-    """Run all tests."""
-    if not os.environ.get('CONGRESS_API_KEY'):
-        print("❌ ERROR: CONGRESS_API_KEY environment variable not set.")
-        print("Please set your Congress API key before running this test.")
-        print("Example: export CONGRESS_API_KEY='your_api_key_here'")
-        return
-    
-    print("🧪 Congress Scraper Semantic Filtering Test Suite")
-    print("=" * 60)
-    print("This test demonstrates the fix for the issue where different")
-    print("queries were returning the same bills from the Congress API.")
-    print()
-    
-    try:
-        # Run all tests
-        test_semantic_filtering()
-        test_broad_interests()
-        test_handler_function()
-        
-        print("\n" + "=" * 60)
-        print("🎉 All tests completed successfully!")
-        print("\nKey improvements demonstrated:")
-        print("✅ URL encoding fixed for multi-word queries")
-        print("✅ Semantic filtering implemented to work around broken Congress API search")
-        print("✅ Different interests now return different, relevant bills")
-        print("✅ Bills are properly matched to their most relevant interests")
-        print("✅ Similarity scores provide transparency in matching")
-        
-    except Exception as e:
-        print(f"\n❌ Test failed with error: {e}")
-        import traceback
-        traceback.print_exc()
-
-if __name__ == "__main__":
-    main()
